requirements.txt
certifi 
discord.py[voice]
py-cpuinfo
Pillow
requests
bs4
typing_extensions
git+https://github.com/cloudwithax/pomice
python-dotenv
imgurpython
psutil